insert a post-update step to allow for bigger changes
The move of the config files and some of the other executable and config files
that were sprinkled all over the place causes some potential issues, depending
how old the version is that the user is upgrading from. Try to address most of
them in this script that gets run as part of restarting adsb-setup (which in
return happens at the end of an update).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>